requirements.txt
pandas>=2.0
openpyxl>=3.1
xlrd==2.0.1
numpy>=1.24
pydantic>=2.0
lxml>=4.9
python-dotenv>=1.0
rich>=13.0
rapidfuzz>=3.0
openai>=1.0
